# core-api/legacy_scripts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Financial Analysis API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "http://localhost:8000",
      "description": "Production server (update with actual URL)"
    }
  ],
  "paths": {
...
    "/api/reports/summary": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Summary Report",
        "description": "Generate a comprehensive financial summary report including:\n    - Income summary with top categories\n    - Expense summary with top categories\n    - Cash flow summary\n    - Financial health indicators\n    \n    The report can be generated in JSON or CSV format.",
        "operationId": "get_summary_report_api_reports_summary_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (default: 30 days ago)",
              "title": "Start Date"
            },
            "description": "Start date (default: 30 days ago)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (default: today)",
              "title": "End Date"
            },
            "description": "End date (default: today)"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/src__financial_analysis__api__schemas__report__AnalysisModeEnum",
              "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all",
              "default": "analysis"
            },
            "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReportFormatEnum",
              "description": "Export format: 'json' or 'csv'",
              "default": "json"
            },
            "description": "Export format: 'json' or 'csv'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/income": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Income Detail Report",
        "description": "Generate a detailed income report showing:\n    - Total income for the period\n    - Transaction count and averages\n    - Complete breakdown by category with percentages\n    \n    Useful for understanding income sources and patterns.",
        "operationId": "get_income_detail_report_api_reports_income_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (default: 30 days ago)",
              "title": "Start Date"
            },
            "description": "Start date (default: 30 days ago)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (default: today)",
              "title": "End Date"
            },
            "description": "End date (default: today)"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/src__financial_analysis__api__schemas__report__AnalysisModeEnum",
              "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all",
              "default": "analysis"
            },
            "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReportFormatEnum",
              "description": "Export format: 'json' or 'csv'",
              "default": "json"
            },
            "description": "Export format: 'json' or 'csv'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeDetailReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/expenses": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Expense Detail Report",
        "description": "Generate a detailed expense report showing:\n    - Total expenses for the period\n    - Transaction count and averages\n    - Complete breakdown by category with percentages\n    \n    Useful for understanding spending patterns and identifying areas for optimization.",
        "operationId": "get_expense_detail_report_api_reports_expenses_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (default: 30 days ago)",
              "title": "Start Date"
            },
            "description": "Start date (default: 30 days ago)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (default: today)",
              "title": "End Date"
            },
            "description": "End date (default: today)"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/src__financial_analysis__api__schemas__report__AnalysisModeEnum",
              "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all",
              "default": "analysis"
            },
            "description": "Analysis mode: 'analysis' excludes transfers, 'complete' includes all"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReportFormatEnum",
              "description": "Export format: 'json' or 'csv'",
              "default": "json"
            },
            "description": "Export format: 'json' or 'csv'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseDetailReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/reconciliation": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Reconciliation Report",
        "description": "Generate a reconciliation report showing ALL transactions including transfers.\n    \n    This report uses COMPLETE mode to show:\n    - All transactions in the period (including transfers, credit card payments, etc.)\n    - Complete transaction details for reconciliation\n    - Summary statistics\n    \n    Useful for:\n    - Bank reconciliation\n    - Verifying all transactions are accounted for\n    - Auditing purposes\n    - Complete financial picture",
        "operationId": "get_reconciliation_report_api_reports_reconciliation_get",
        "parameters": [
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date (default: 30 days ago)",
              "title": "Start Date"
            },
            "description": "Start date (default: 30 days ago)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "End date (default: today)",
              "title": "End Date"
            },
            "description": "End date (default: today)"
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReportFormatEnum",
              "description": "Export format: 'json' or 'csv'",
              "default": "json"
            },
            "description": "Export format: 'json' or 'csv'"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconciliationReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashflowSummary": {
        "properties": {
          "net_cashflow": {
            "type": "number",
            "title": "Net Cashflow",
            "description": "Net cash flow (income - expenses)"
          },
          "total_income": {
            "type": "number",
            "title": "Total Income",
            "description": "Total income"
          },
          "total_expenses": {
            "type": "number",
            "title": "Total Expenses",
            "description": "Total expenses"
          }
        },
        "type": "object",
        "required": [
          "net_cashflow",
          "total_income",
          "total_expenses"
        ],
        "title": "CashflowSummary",
        "description": "Cash flow summary in a report."
      },
...
      "CategorySummary": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Category name"
          },
          "amount": {
            "type": "number",
            "title": "Amount",
            "description": "Total amount for the category"
          },
          "percentage": {
            "type": "number",
            "title": "Percentage",
            "description": "Percentage of total"
          }
        },
        "type": "object",
        "required": [
          "category",
          "amount",
          "percentage"
        ],
        "title": "CategorySummary",
        "description": "Category summary in a report."
      },
...
      "ExpenseDetailReportResponse": {
        "properties": {
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "Type of report"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod",
            "description": "Report period"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Analysis mode used"
          },
          "total_expenses": {
            "type": "number",
            "title": "Total Expenses",
            "description": "Total expenses"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "description": "Number of transactions"
          },
          "average_transaction": {
            "type": "number",
            "title": "Average Transaction",
            "description": "Average transaction amount"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryDetail"
            },
            "type": "array",
            "title": "Categories",
            "description": "Expenses by category"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "period",
          "mode",
          "total_expenses",
          "transaction_count",
          "average_transaction",
          "categories"
        ],
        "title": "ExpenseDetailReportResponse",
        "description": "Response for detailed expense report."
      },
...
      "ExpenseSummary": {
        "properties": {
          "total": {
            "type": "number",
            "title": "Total",
            "description": "Total expenses"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "description": "Number of expense transactions"
          },
          "average_transaction": {
            "type": "number",
            "title": "Average Transaction",
            "description": "Average expense transaction"
          },
          "top_categories": {
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "type": "array",
            "title": "Top Categories",
            "description": "Top 5 expense categories"
          }
        },
        "type": "object",
        "required": [
          "total",
          "transaction_count",
          "average_transaction",
          "top_categories"
        ],
        "title": "ExpenseSummary",
        "description": "Expense summary in a report."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthIndicators": {
        "properties": {
          "income_to_expense_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Income To Expense Ratio",
            "description": "Income to expense ratio"
          },
          "savings_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Savings Rate",
            "description": "Savings rate as percentage"
          },
          "average_daily_income": {
            "type": "number",
            "title": "Average Daily Income",
            "description": "Average daily income"
          },
          "average_daily_expense": {
            "type": "number",
            "title": "Average Daily Expense",
            "description": "Average daily expense"
          },
          "average_daily_cashflow": {
            "type": "number",
            "title": "Average Daily Cashflow",
            "description": "Average daily cash flow"
          }
        },
        "type": "object",
        "required": [
          "average_daily_income",
          "average_daily_expense",
          "average_daily_cashflow"
        ],
        "title": "HealthIndicators",
        "description": "Financial health indicators."
      },
...
      "IncomeDetailReportResponse": {
        "properties": {
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "Type of report"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod",
            "description": "Report period"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Analysis mode used"
          },
          "total_income": {
            "type": "number",
            "title": "Total Income",
            "description": "Total income"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "description": "Number of transactions"
          },
          "average_transaction": {
            "type": "number",
            "title": "Average Transaction",
            "description": "Average transaction amount"
          },
          "categories": {
            "items": {
              "$ref": "#/components/schemas/CategoryDetail"
            },
            "type": "array",
            "title": "Categories",
            "description": "Income by category"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "period",
          "mode",
          "total_income",
          "transaction_count",
          "average_transaction",
          "categories"
        ],
        "title": "IncomeDetailReportResponse",
        "description": "Response for detailed income report."
      },
...
      "IncomeSummary": {
        "properties": {
          "total": {
            "type": "number",
            "title": "Total",
            "description": "Total income"
          },
          "transaction_count": {
            "type": "integer",
            "title": "Transaction Count",
            "description": "Number of income transactions"
          },
          "average_transaction": {
            "type": "number",
            "title": "Average Transaction",
            "description": "Average income transaction"
          },
          "top_categories": {
            "items": {
              "$ref": "#/components/schemas/CategorySummary"
            },
            "type": "array",
            "title": "Top Categories",
            "description": "Top 5 income categories"
          }
        },
        "type": "object",
        "required": [
          "total",
          "transaction_count",
          "average_transaction",
          "top_categories"
        ],
        "title": "IncomeSummary",
        "description": "Income summary in a report."
      },
...
      "ReconciliationReportResponse": {
        "properties": {
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "Type of report"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod",
            "description": "Report period"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Analysis mode (always 'complete' for reconciliation)"
          },
          "summary": {
            "$ref": "#/components/schemas/ReconciliationSummary",
            "description": "Summary statistics"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionDetail"
            },
            "type": "array",
            "title": "Transactions",
            "description": "All transactions in period"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "period",
          "mode",
          "summary",
          "transactions"
        ],
        "title": "ReconciliationReportResponse",
        "description": "Response for reconciliation report.",
        "example": {
          "mode": "complete",
          "period": {
            "days": 31,
            "end_date": "2025-01-31",
            "start_date": "2025-01-01"
          },
          "report_type": "reconciliation",
          "summary": {
            "net_cashflow": 2300.0,
            "total_expenses": 3200.0,
            "total_income": 5500.0,
            "transaction_count": 75
          },
          "transactions": [
            {
              "amount": 5000.0,
              "category": "Income",
              "classification": "Standard Transaction",
              "date": "2025-01-31",
              "description": "Salary",
              "source": "Manual Entry",
              "type": "Income"
            }
          ]
        }
      },
...
      "ReportFormatEnum": {
        "type": "string",
        "enum": [
          "json",
          "csv"
        ],
        "title": "ReportFormatEnum",
        "description": "Report export formats."
      },
      "ReportPeriod": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "Start date of the period"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "End date of the period"
          },
          "days": {
            "type": "integer",
            "title": "Days",
            "description": "Number of days in the period"
          }
        },
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "days"
        ],
        "title": "ReportPeriod",
        "description": "Report period information."
      },
...
      "SummaryReportResponse": {
        "properties": {
          "report_type": {
            "type": "string",
            "title": "Report Type",
            "description": "Type of report"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriod",
            "description": "Report period"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "Analysis mode used"
          },
          "income": {
            "$ref": "#/components/schemas/IncomeSummary",
            "description": "Income summary"
          },
          "expenses": {
            "$ref": "#/components/schemas/ExpenseSummary",
            "description": "Expense summary"
          },
          "cashflow": {
            "$ref": "#/components/schemas/CashflowSummary",
            "description": "Cash flow summary"
          },
          "health_indicators": {
            "$ref": "#/components/schemas/HealthIndicators",
            "description": "Financial health indicators"
          }
        },
        "type": "object",
        "required": [
          "report_type",
          "period",
          "mode",
          "income",
          "expenses",
          "cashflow",
          "health_indicators"
        ],
        "title": "SummaryReportResponse",
        "description": "Response for summary report.",
        "example": {
          "cashflow": {
            "net_cashflow": 2000.0,
            "total_expenses": 3000.0,
            "total_income": 5000.0
          },
          "expenses": {
            "average_transaction": 60.0,
            "top_categories": [
              {
                "amount": 800.0,
                "category": "Groceries",
                "percentage": 26.67
              }
            ],
            "total": 3000.0,
            "transaction_count": 50
          },
          "health_indicators": {
            "average_daily_cashflow": 64.52,
            "average_daily_expense": 96.77,
            "average_daily_income": 161.29,
            "income_to_expense_ratio": 1.67,
            "savings_rate": 40.0
          },
          "income": {
            "average_transaction": 500.0,
            "top_categories": [
              {
                "amount": 4500.0,
                "category": "Salary",
                "percentage": 90.0
              }
            ],
            "total": 5000.0,
            "transaction_count": 10
          },
          "mode": "analysis",
          "period": {
            "days": 31,
            "end_date": "2025-01-31",
            "start_date": "2025-01-01"
          },
          "report_type": "summary"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "src__financial_analysis__api__schemas__report__AnalysisModeEnum": {
        "type": "string",
        "enum": [
          "analysis",
          "complete"
        ],
        "title": "AnalysisModeEnum",
        "description": "Analysis mode for reports."
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (to be implemented)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication (to be implemented)"
      }
    },
...
  },
...
}